requirements.txt
Keras
Keras-Applications
Keras-Preprocessing
numpy
Pillow==6.2.0
scipy
seaborn
tensorboard
tensorflow
tensorflow-estimator
jupyter
